+++ data-browser/src/components/HotKeyWrapper.tsx
@@ -71,7 +71,7 @@ function HotKeysWrapper({ children }: Props): JSX.Element {
     shortcuts.edit,
     e => {
       e.preventDefault();
-      Client.isValidURL(subject) && navigate(editURL(subject!));
+      Client.isValidSubject(subject) && navigate(editURL(subject!));
     },
     {},
     [subject],
@@ -80,7 +80,7 @@ function HotKeysWrapper({ children }: Props): JSX.Element {
     shortcuts.data,
     e => {
       e.preventDefault();
-      Client.isValidURL(subject) && navigate(dataURL(subject!));
+      Client.isValidSubject(subject) && navigate(dataURL(subject!));
     },
     {},
     [subject],